create event without guest, event review, pricing, package web, payment status, attendee tracker, modified controller and table creation [heart and symond task]

+++ Backend/database/migrations/2024_10_25_093821_create_events_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('type');
             $table->integer('pax');
             $table->string('status');
-            $table->integer('totalPrice')->nullable();
+            $table->decimal('totalPrice', 10, 2)->nullable();
             $table->date('date');
             $table->time('time');
             $table->string('location');
